split secrets, credentials from configuration
