+++ lega_init/deploy.py
@@ -1,7 +1,10 @@
 import logging
 import click
 import sys
+import os
+import errno
 from .configure import ConfigGenerator
+from .key_generation import SecurityConfigGenerator
 from pathlib import Path
 from ruamel.yaml.scalarstring import DoubleQuotedScalarString as dq
 
@@ -18,31 +21,45 @@ def create_config(_localega, config_path, cega):
     _here = Path(config_path)
     config_dir = _here
 
+    if not os.path.exists(config_dir):
+        try:
+            os.makedirs(config_dir)
+        except OSError as exc:  # Guard against race condition
+            if exc.errno != errno.EEXIST:
+                raise
+
     # Generate Configuration
-    conf = ConfigGenerator(config_dir, _localega['key']['name'], _localega['email'])
-    conf.generate_mq_config()
+    token_payload = {"iss": "http://data.epouta.csc.fi",
+                     "authorities": ["EGAD01"]}
+    sec_config = SecurityConfigGenerator(config_dir, _localega['key']['name'], _localega['email'])
+    token_keys = sec_config.generate_token(_localega['keys_password'])
+    pgp_passphrase = sec_config._generate_secret(32)
+    cega_mq_auth_secret = sec_config._generate_secret(32)
+    mq_auth_secret = sec_config._generate_secret(32)
+    sec_config.generate_ssl_certs(country=_localega['ssl']['country'], country_code=_localega['ssl']['country_code'],
+                                  location=_localega['ssl']['location'], org=_localega['ssl']['org'], email=_localega['email'])
+    pgp_pair = sec_config.generate_pgp_pair(comment=_localega['key']['comment'],
+                                            passphrase=pgp_passphrase, armor=True, active=True)
+    auth_keys = sec_config.generate_user_auth(_localega['keys_password'])
+    conf = ConfigGenerator(config_path, token_keys, auth_keys, pgp_pair)
+    conf.generate_mq_config(mq_auth_secret)
     if cega:
-        conf.generate_cega_mq_auth()
+        conf.generate_cega_mq_auth(cega_mq_auth_secret)
         conf.generate_user_auth(_localega['keys_password'])
-        conf._trace_secrets.update(cega_users_pass=dq(conf._generate_secret(32)))
-    conf.generate_token(_localega['keys_password'])
-    pg_in_password = conf._generate_secret(32)
-    pg_out_password = conf._generate_secret(32)
+        conf._trace_secrets.update(cega_users_pass=dq(sec_config._generate_secret(32)))
+    pg_in_password = sec_config._generate_secret(32)
+    pg_out_password = sec_config._generate_secret(32)
     conf._trace_secrets.update(pg_in_password=dq(pg_in_password))
     conf._trace_secrets.update(pg_out_password=dq(pg_out_password))
-    s3_access_key = conf._generate_secret(16)
+    s3_access_key = sec_config._generate_secret(16)
     conf._trace_secrets.update(s3_access_key=dq(s3_access_key))
-    s3_secret_key = conf._generate_secret(32)
+    s3_secret_key = sec_config._generate_secret(32)
     conf._trace_secrets.update(s3_secret_key=dq(s3_secret_key))
-    shared_pgp_password = conf._generate_secret(32)
+    shared_pgp_password = sec_config._generate_secret(32)
     conf._trace_secrets.update(shared_pgp_password=dq(shared_pgp_password))
-    pgp_passphrase = conf._generate_secret(32)
     conf._trace_secrets.update(pgp_passphrase=dq(pgp_passphrase))
-
-    conf.add_conf_key(_localega['key']['expire'], _localega['key']['id'], comment=_localega['key']['comment'],
-                      passphrase=pgp_passphrase, armor=True, active=True)
-    conf.generate_ssl_certs(country=_localega['ssl']['country'], country_code=_localega['ssl']['country_code'],
-                            location=_localega['ssl']['location'], org=_localega['ssl']['org'], email=_localega['email'])
+    conf.generate_token(token_payload)
+    conf.add_conf_key(_localega['key']['id'], armor=True)
 
     conf.write_trace_yml()
